% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owintosppolygons.R
\name{as.Polygon.owin}
\alias{as.Polygon.owin}
\alias{as.Polygons.Polygon}
\alias{as.spdf.owin}
\title{convert owin polygons to an sp SpatialPolygonsDataFrame}
\usage{
as.Polygon.owin(owinpoly)

as.Polygons.Polygon(sppolygon, id = "NA")

as.spdf.owin(owinpolys, auxdata, idcolumn = NULL,
  proj4string = CRS(as.character(NA)), match.ID = TRUE)
}
\arguments{
\item{owinpoly}{A single spatstat owin object of type polygonal or rectangular}

\item{sppolygon}{is a "Polygon" object for sp.}

\item{id}{ids to be assigned to the polygon(s). Default is NA}

\item{auxdata}{dataframe of auxilary information for each polygon, must include an id column. Must have rows corresponding to each polygon.}

\item{idcolumn}{a character refering to the column to use for ids in auxdata. If none is supplied then uses rownames}

\item{proj4string}{Projection information see sp and rgdal for more detail}

\item{match.ID}{see help for \code{SpatialPolygonsDataFrame}. If TRUE check that row names of auxdata match to id possibly with rearrangement
If character match polygon ids with a particular column in auxdata}

\item{owinpoly}{is a list of owin polygons}
}
\value{
\code{as.Polygon.owin} returns a single sp polygon object

an sp "Polygons" object from a single Polygon object and an id
}
\description{
functions for converting an owin polygon or rectangle to the sp polygon formats. This is useful for interfacing with rgdal and raster
}
\details{
\code{as.Polygon.owin} for converting owin objects to sp polygons without an projection information
}
\section{Functions}{
\itemize{
\item \code{as.Polygons.Polygon}: From a single sp \code{Polygon} object returns an sp \code{Polygons} object

\item \code{as.spdf.owin}: Returns a SpatialPolgyonsDataFrame given a list of owin polygons, ids and projection information
}}
\examples{
data(polygontest)
tess <- quadrats(polygontest)
tilelist <- tiles(tess)


}
\seealso{
SpatialPolygonsDataFrame
}

