% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagPixel.R
\name{contagpixelgrid}
\alias{adjacency}
\alias{contagpixelgrid}
\title{Pixel Contagion}
\usage{
contagpixelgrid(xi, w)

adjacency(xi, w)
}
\arguments{
\item{xi}{an owin mask representing a realisation of RACS}

\item{w}{the window of observation}
}
\description{
Function for calculating the standard LPI contagion from a pixel grid.
Refer to FRAGSTATS for a full description of contagion.
}
\details{
The `double count' method as described in the FRAGSTATS manual on adjacency matrix.
Considering not xi as another phase and using a 4-neighbourhood.
}
\section{Functions}{
\itemize{
\item \code{adjacency}: Calculates the adjacency matrix used in the pixel contagion
}}
\section{Warning}{
 Will fail if there are no adjacencies
}

