% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagSphCont.R
\name{contagSphCont}
\alias{contagSphCont}
\title{Spherical Contact version of Contagion}
\usage{
contagSphCont(xiH, xiHc, p, normalise = FALSE)
}
\arguments{
\item{xiH}{Spherical contact function for xi.}

\item{xiHc}{Spherical contact distribution for the complement of xi.}

\item{p}{An estimate of the coverage fraction of a RACS \eqn{\Xi}.
The sample points should be same as xiH for now.}

\item{normalise}{Optional. If true normalises the results so that all RACS return a value between}
}
\description{
Calculates a spherical contact version of contagion. 
This version of contagion gauges closeness/mixing between \eqn{\Xi} and \eqn{\Xi^c}
 using the spherical contact distribution. 
 Given a distance \eqn{r} and an arbitrary point \eqn{x}
 in \eqn{\Xi} then \eqn{\Xi} close means everything with \eqn{r} of \eqn{x} is in \eqn{\Xi}.
 \eqn{\Xi^c} close to \eqn{x} means there is at least some part of \eqn{\Xi^c} within \eqn{r} of \eqn{x}.
}
\details{
xiH should be a function of radius that estimates the probability 
\deqn{xiH(r)\approx P(B_r(x) \subseteq \Xi^c)}
 of a disc around an arbitrary point \eqn{x} is contained in \eqn{\Xi^c.}
Similary xiHc should be the probability of disc being fully contained in \eqn{\Xi}
\deqn{xiHc(r)\approx P(B_r(o) \subseteq \Xi).}

If \code{normalise} is \code{TRUE} then divides by 
\eqn{\frac{-4}{e}ln(\frac{1}{e})} and adds 1 so normalised spherical contact contagion is
\deqn{
1+(\frac{-4}{e}ln(\frac{1}{e}))^{-1} \text{unnormalised contagion}
}
This makes contagion vary between 0 and 1 for all 2 phase processes.
}

