% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdatatoram.R
\name{getrastervaluesofpolygons}
\alias{getrastervaluesofpolygons}
\alias{putencompassingrastervaluesinram}
\alias{transfertoram}
\title{Functions for reading loading remote sensing data into RAM for quick extraction later}
\usage{
getrastervaluesofpolygons(polygons, inrasterfilename)

putencompassingrastervaluesinram(polygons, inrasterfilename)

transfertoram(raster, extent, sizefactor)
}
\arguments{
\item{polygons}{is a SpatialPolygonDataFrame.}

\item{inrasterfilename}{is the location of a raster file that contains *all* of the polygons}

\item{raster}{Any raster object.}

\item{extent}{Giving the extent of values to load into ram}

\item{sizefactor}{A multiplicative safety margin - it must be possible to hold \code{sizefactor}*raster data size in RAM.}
}
\description{
When extracting multiple polygons of data it is much faster to read in a large chunk of data to ram and then extract the polygons from this chunk then to extract each polygon separately from the data on HDD. 
These functions are designed to help with this. Currently using raster functions, I think the best thing will be to shift to rgdal read functions in the future.
}
\section{Functions}{
\itemize{
\item \code{putencompassingrastervaluesinram}: 

\item \code{transfertoram}: Tests if a chunk of data can be fit in ram, and if so transfers it to ram.
}}

