% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelsmooth.R
\name{kernelsmooth}
\alias{kernelsmooth}
\title{Kernel Smoothing Tools}
\usage{
kernelsmooth(im, bandwidth, kernel = "Epanechnikov")
}
\arguments{
\item{im}{an image for smoothing.}

\item{bandwidth}{Bandwidth for the chosen kernel (in reciprocal of the units of \code{im})}

\item{kernel}{Specifies the kernel to use. Currently only Epanechnikov is supported (and is the default).}
}
\value{
A smoothed image in spatstat \code{im} format
}
\description{
Functions for kernel smoothing spatstat images. This is mostly for smoothing spectral density estimates.
Only kernel currently supported is Epanechnikov.
}
\details{
Uses FFT to do the calculation.
}

