% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laslettTransform.R
\name{lasletttransform}
\alias{findlowlefttangentpts}
\alias{lasletttransform}
\title{Laslett's Transform}
\usage{
lasletttransform(xi)

findlowlefttangentpts(xi)
}
\arguments{
\item{xi}{An owin mask object.}
}
\value{
\code{lasletttranform} Returns a ppp of the tranformed location of tangent points with transformed window.
       \code{findlowlefttangentpts} returns two lists, labelled X and Y respectively of the X and Y coordinates of tangent point.
}
\description{
Performs Laslett's tranform on a raster map. \code{findlowlefttangentpts} is function used to find the tangent points.
See ** reference.
}
\section{Functions}{
\itemize{
\item \code{findlowlefttangentpts}: Returns a list of X and Y coordinates of tangent points
}}

