% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgermgrain.R
\name{rgermgrain}
\alias{placegrainsfromlib}
\alias{rgermgrain}
\title{Functions for flexibly generating a simulation of germ grain model}
\usage{
placegrainsfromlib(pp, grainlib, replace = TRUE, prob = NULL)
}
\arguments{
\item{pp}{A point pattern (in \code{ppp} format).}

\item{grainlib}{A list (aka. library/population) of grains (in \code{solist} format) that grains will be selected from}

\item{replace}{passed directly to \code{\link[base]{sample}}. When TRUE grains are chosen from library with replacement.}

\item{prob}{A list of probability weights for each grain in the library. Passed directly to \code{\link[base]{sample}}. If NULL I'm pretty sure grains all have equal probability.}
}
\value{
Returns an \code{owin} object.
}
\description{
Use point process simulations from spatstat to generate a point process in desired region (must include a buffer). \code{placegrainsfromlib} can then be used to randomly select grains from a library and place them around each point. Crop out the buffer zone to get a simulation.
}
\details{
\code{placegrainsfromlib} randomly samples from a library of grains (\code{grainlib}) and places these on the points in \code{pp}.
}
\author{
Kassel Hingee
}

