% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraldensity.R
\name{spectraldensity}
\alias{kernelsmooth}
\alias{spectraldensity}
\alias{unsmoothedspectraldensity}
\title{Spectral Density of a RACS}
\usage{
spectraldensity(Xi, w, bandwidth, suffspecres = NULL,
  kernel = "Epanechnikov")

kernelsmooth(im, bandwidth, kernel = "Epanechnikov")

unsmoothedspectraldensity(Xi, w, suffspecres = NULL)
}
\arguments{
\item{Xi}{A rectangular observation of \eqn{Xi}. NA's are assumed to mean outside \eqn{\Xi} rather than missing data. Xi must be pixel mask owin object. (**I haven't assesed the theory/computations for non-rectangular windows but its probably the same)}

\item{w}{Observation window (must be rectangular for now)}

\item{bandwidth}{Bandwidth for the chosen kernel (in reciprocal of the units of \code{Xi})}

\item{suffspecres}{Optional. A desired spectral resolution, the output resolution will be smaller than \code{suffspecres}
If it is not provided then resolution will be 2*pi/(size of spatial extent of window), for example a window 500m x 500m will result in spectral resolution of 2*pi/500 in both X and Y directions.
Can be of length 1 (applies to all dimensions), or length 2 (a differen desired resolution for each dimension).}

\item{kernel}{Specifies the kernel to use. Currently only Epanechnikov is supported (and is the default).}
}
\description{
\code{unsmoothedspectraldensity} estimates the spectral density of a RACS without any kernal smoothing.
\code{spectraldensity} smoothes the estimate and is my recommended function due to known asymptotoic properties [1].
}
\details{
Applies FFT to the input image and takes the square of the magnitude to estimate the (non-kernel smoothed) spectral density
as described in [B\"{o}hm 2004] (the very first equation or equivalently \eqn{\hat{h}(x)} of equation 4.6 in the 2002 report).  

\eqn{\hat{h}(z) := \frac{1}{|w|}\left|\int_{R^2}1_w(x) (1_\Xi(x)-p) dx\right|}

This is what \code{unsmoothedspectraldensity} returns.
 \code{spectraldensity} then kernel smooths \eqn{\hat{h}} before returning a spectral density estimate.

\code{kernelsmooth} smooths an image using a kernel. Currently the only kernel availabe is Epanechnikov.
}
\section{WARNING}{
 WARNING RESULTS OF THIS FUNCTION HAVENT BEEN TESTED - THEY COULD BE WILDLY WRONG

Also final result should have odd pixel dimensions (so that the spectral density at origin is easy to pull out), but convolve.im returns even dimensions, I don't know why yet.
}
\examples{
#apply to heather
xi <- heather$coarse
specdens <- spectraldensity(xi,Frame(xi),20)
plot(specdens)

#See what unsmoothed version looks like
unsmspecdens <- unsmoothedspectraldensity(xi,Frame(xi))
plot(unsmspecdens)
plot(solist(smoothed = specdens,unsmoothed = unsmspecdens),
       main = "Spectral Density Estimates",
       equal.scales=TRUE)

}
\references{
B\"{o}hm, S., Heinrich, L., Schmidt, V., 2004. Kernel Estimation of the Spectral Density of Stationary Random Closed Sets. Australian & New Zealand Journal of Statistics 46, 41--51. doi:10.1111/j.1467-842X.2004.00310.x
}

