% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageprobabilityVarianceEst.R
\name{varCovProb}
\alias{asympvarP}
\alias{varCovProb}
\title{Variances and Confidence Intervals for Coverage Fraction Estimates}
\usage{
varCovProb(Xi, w)
}
\arguments{
\item{Xi}{is an observation (in owin) format of a RACS.}

\item{w}{is the corresponding observation window.}
}
\value{
varcovProb returns a variance estimate using the covariance function and an estimate of area fraction). This will not be as good as the spectral density based estimate**.
asympvarP() isn't constructed yet but will estimate a variance assuming that things are close to Gaussian distributions.
}
\description{
Functions for estimating variance (and confidence intervals)  of the coverage fraction estimates
}
\details{
Molchanov (1997 Ch3.1) notes that the variance of the estimator \code{\link[stationaryracsinference]{covpest}} is 
\deqn{var = \frac{1}{A(w)^2}\int_W \gamma_W(v)C(v) - p^2 dv}

This function assumes that the probability of two points being in \eqn{\Xi} is independent \deqn{C(v) = p^2} outside the calculated window of covariance which is...
}
\examples{
#**To come later** 
}

