% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmanyfv.R
\name{manylines.fv}
\alias{manylines.fv}
\title{Plot multiple fv objects of similar ilk}
\usage{
manylines.fv(fvlist, xname, yname, ..., add = FALSE, xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{fvlist}{A list of fv objects}

\item{xname}{the name of the value to associate with the x-axis}

\item{yname}{the name of the value to associate with the y-axis}

\item{...}{Plotting parameters to be passed to \code{title} and \code{lines}}

\item{add}{If true will add lines to current plot, otherwise will start new plot and will require xlim and ylim.}

\item{xlim}{Plotting limits of the horizontal direction}

\item{ylim}{Plotting limits of the vertical direction}
}
\value{
List of return values from lines (typically a list of NULL values)
}
\description{
Plots multiple fv objects on the same window.
}
\examples{
hestf <- Hest(heather$fine, W =Frame(heather$fine))
hestm <- Hest(heather$medium, W =Frame(heather$medium))
hestc <- Hest(heather$coarse, W =Frame(heather$coarse))
fvlist <- list(hestf,hestm,hestc)
names(fvlist) <- c("fine", "medium", "coarse")
manylines.fv(fvlist, xname="r", yname="km", xlim=c(0,0.35), ylim=c(0,1))

## Fancier plot
out <- manylines.fv(fvlist, xname="r", yname="km", xlim=c(0,0.35), ylim=c(0,1), 
lwd=c(1,2,3), main="Spherical Contact Distribution\\n from different resolutions", col=rainbow(length(fvlist)))
}

