% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbooldeterministicdiscs.R
\name{simulateBooleanDetermDiscs}
\alias{booldetermdiscs_truecoveragefrac}
\alias{simulateBooleanDetermDiscs}
\title{Simulation of Boolean Model of Deterministic Discs}
\usage{
simulateBooleanDetermDiscs(lambda, discr, window)

booldetermdiscs_truecoveragefrac(lambda, discr)
}
\arguments{
\item{lambda}{Intensity of the germ process (which is a Poisson point process)}

\item{discr}{Radius of the discs}

\item{window}{The window to simulate in (an owin object)}
}
\value{
\code{simulateBooleanDetermDiscs} returns an owin object containing the locations covered by the Boolean model. The window information is not contained.

\code{booldetermdiscs_truecoveragefrac} returns the true coverage fraction given the intensity and the radius.
}
\description{
Function for simulating a Boolean model with deterministic discs, and also functions for calculating theoretical properties such as coverage function, covariance, and spectral density.
}

